Refactor float builtins that take an int argument
This generifies some of the signatures that took only raw pointers,
allowing them to support `multi_ptr`s. This also changes the SSCP
signatures of all of the functions, as the SSCP form associated
32-bit ints with float and 64-bit ints with double - change them both
to 32-bit ints for consistency and code simplification. (We were really
just converting back and forth for no reason).